# Conflicts:
#	asset/asset-web/resources/javascript/common/common.js

+++ hr/hr-leave/src/main/java/org/egov/eis/web/contract/EligibleLeave.java
@@ -57,16 +57,16 @@ import lombok.ToString;
 @ToString
 public class EligibleLeave {
 
-	@JsonProperty("leaveType")
-	private Long leaveType;
+    @JsonProperty("leaveType")
+    private Long leaveType;
 
-	@JsonProperty("employee")
-	private Long employee;
+    @JsonProperty("employee")
+    private Long employee;
 
-	@JsonProperty("noOfDays")
-	private Float noOfDays;
+    @JsonProperty("noOfDays")
+    private Double noOfDays;
 
-	@JsonProperty("asOnDate")
-	private String asOnDate;
+    @JsonProperty("asOnDate")
+    private String asOnDate;
 
 }
